backend/requirements.txt
fastapi==0.92.0
uvicorn==0.17.6
sqlalchemy==1.4.42
python-dotenv==0.20.0
alembic==1.9.1
psycopg2-binary==2.9.5
python-multipart==0.0.5
bcrypt==3.2.0